requirement.txt
py-cord
python-dotenv
pocketbase